requirements.txt
alembic==1.13.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.36.0
Deprecated==1.2.14
docker==7.1.0
efficientnet_pytorch==0.7.1
einops==0.8.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
graphene==3.4.1
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.1
wrapt==1.16.0
zipp==3.20.2